Splits tests into separate, descriptive files
