+++ public_officials/unit_tests/test_legislator_model.py
@@ -6,90 +6,6 @@ from public_officials.models import *
 from django.contrib.auth.models import User
 import vcr
 
-class HomePageTest(TestCase):
-    def test_root_url_resolves_to_home_page_view(self):
-        found = resolve('/')
-        self.assertEqual(found.func, home_page)
-
-    def test_home_page_returns_correct_template(self):
-        response = self.client.get('/')
-        self.assertTemplateUsed(response, 'public-officials/home.html')
-
-class StatePageTest(TestCase):
-    def test_state_url_resolves_to_state_view(self):
-        found = resolve('/state/')
-        self.assertEqual(found.func, state_legislators)
-
-    def test_state_legislators_returns_correct_template(self):
-        response = self.client.get('/state/?state=Texas')
-        self.assertTemplateUsed(response, 'public-officials/state.html')
-
-class DetailPageTest(TestCase):
-    def test_detail_url_resolves_to_legislator_detail_view(self):
-        found = resolve('/legislators/1/')
-        self.assertEqual(found.func, legislator_detail)
-
-    def test_legislator_detail_page_returns_correct_template(self):
-        with vcr.use_cassette('cassettes/get_complete_legislator_profile'):
-            legislator = Legislator.objects.create(first_name="Diana",
-                                                   last_name="Degette",
-                                                   phone="12345",
-                                                   email="email@email.com",
-                                                   state_name="Colorado",
-                                                   pid="D000197",
-                                                   chamber="house",
-                                                   term_start="2017-01-03",
-                                                   term_end="2019-01-03",
-                                                   party="D",
-                                                   state="CO",
-                                                   cid="N00006134")
-
-            response = self.client.get('/legislators/%d/' % legislator.id)
-            self.assertTemplateUsed(response, 'public-officials/detail.html')
-
-class SenatorPageTest(TestCase):
-    def test_senators_url_resolves_to_senators_index(self):
-        found = resolve('/senators/')
-        self.assertEqual(found.func, senator_index)
-
-    def test_senator_index_page_returns_correct_template(self):
-        Legislator.objects.create(first_name="Diana",
-                                  last_name="Degette",
-                                  phone="12345",
-                                  email="email@email.com",
-                                  state_name="Colorado",
-                                  pid="D000197",
-                                  chamber="senate",
-                                  term_start="2017-01-03",
-                                  term_end="2019-01-03",
-                                  party="D",
-                                  state="CO",
-                                  cid="N00006134")
-        response = self.client.get('/senators/')
-        self.assertTemplateUsed(response, 'public-officials/senators/index.html')
-
-class RepresentativePageTest(TestCase):
-    def test_senators_url_resolves_to_senators_index(self):
-        found = resolve('/representatives/')
-        self.assertEqual(found.func, representative_index)
-
-    def test_senator_index_page_returns_correct_template(self):
-        Legislator.objects.create(first_name="Diana",
-                                  last_name="Degette",
-                                  phone="12345",
-                                  email="email@email.com",
-                                  state_name="Colorado",
-                                  pid="D000197",
-                                  chamber="house",
-                                  term_start="2017-01-03",
-                                  term_end="2019-01-03",
-                                  party="D",
-                                  state="CO",
-                                  cid="N00006134")
-        response = self.client.get('/representatives/')
-        self.assertTemplateUsed(response, 'public-officials/representatives/index.html')
-
-
 class LegislatorModelTest(TestCase):
     def test_a_legislators_attributes(self):
         legislator = Legislator()
@@ -278,51 +194,3 @@ class LegislatorModelTest(TestCase):
         self.assertEqual(2, len(returned_states))
         self.assertEqual('Colorado', returned_states[0])
         self.assertEqual('Texas', returned_states[1])
-
-class RecentBillsPageTest(TestCase):
-    def test_recent_bills_url_resolves_to_bills_index_view(self):
-        found = resolve('/recent-bills/')
-        self.assertEqual(found.func, bills_index)
-
-    def test_recent_bills_url_renders_correct_template(self):
-        response = self.client.get('/recent-bills/')
-        self.assertTemplateUsed(response, 'public-officials/bills/index.html')
-
-class LegislatorRelationshipsTest(TestCase):
-    def test_legislator_has_many_to_many_relationship_with_users(self):
-        legislator = Legislator.objects.create(first_name="Diana",
-                                               last_name="DeGette",
-                                               phone="12345",
-                                               email="email@email.com",
-                                               state_name="Colorado",
-                                               pid="D000197",
-                                               chamber="house",
-                                               term_start="2017-01-03",
-                                               term_end="2019-01-03",
-                                               party="D",
-                                               state="CO",
-                                               cid="N00006134")
-        legislator_two = Legislator.objects.create(first_name="Henry",
-                                                   last_name="Schmojo",
-                                                   phone="12345",
-                                                   email="email@email.com",
-                                                   state_name="Texas",
-                                                   pid="D000197",
-                                                   chamber="house",
-                                                   term_start="2017-01-03",
-                                                   term_end="2019-01-03",
-                                                   party="D",
-                                                   state="TX",
-                                                   cid="N00006134")
-        user = User.objects.create(username='test')
-        user2 = User.objects.create(username='test2')
-
-        legislator.users.add(user)
-        legislator.users.add(user2)
-        legislator_two.users.add(user)
-
-        self.assertEqual(legislator.users.count(), 2)
-        self.assertEqual(legislator_two.users.count(), 1)
-        self.assertEqual(user.legislator_set.all().count(), 2)
-        self.assertEqual(user2.legislator_set.all().count(), 1)
-        self.assertEqual(user2.legislator_set.all().first(), legislator)
