+++ will_legislate_for_money/urls.py
@@ -26,7 +26,6 @@ urlpatterns = [
     url(r'^senators/$', public_official_views.senator_index, name="senator_index"),
     url(r'^representatives/$', public_official_views.representative_index, name="representative_index"),
     url(r'^recent-bills/$', public_official_views.bills_index, name="bills_index"),
-    url(r'^contributions/$', public_official_views.contributions, name="contributions"),
     url(r'^legislators/', include('public_officials.urls')),
     url(r'^admin/', admin.site.urls),
     url(r'^logout/$', public_official_views.log_out, name="log_out"),
